refactor(collector): extract hash.go from base.go
Move ComputeMetricsHash, normalizeValue, and getSortedKeys into
their own file for change-detection hashing.

Part of #75.
